refactor: view 패키지로 변경 및 각 클래스 이름 변경
